requirements.txt
pandas
python-dotenv
openai
PyMuPDF